feat: use answer related hooks
